requirements/requirements.txt
# requirements for the production project.

ipfs-cid==1.0.0
PyYAML==6.0.3
python-magic-bin==0.4.14; sys_platform == 'windows'
python-magic==0.4.27
unf==0.11.0
